requirements.txt
fastapi==0.94.1
uvicorn==0.21.0
SQLAlchemy==2.0.6
passlib==1.7.4
mysqlclient==2.1.1
python-jose==3.3.0
python-multipart==0.0.6
alembic==1.10.2
requests==2.28.2
python-dotenv==1.0.0
